feature/160-Radar-PatientUser-class-create-id-col
160 - Radar PatientUser class is creating an id column due to inheriting
from BaseModel.
